refactor: reuse OptionList Init
